fifth commit
Added brightness control .
i have failed to implement using Setting.System.
for change brightness of the system.
The app keeps shutting down.
but I found a way to adjust the brightness of the app and applied it.
using windowmanager.
I wanted to test, cannot test using AVD.
but I received the value and confirmed it.