tenth commit!
Changed the text font.
Create a dialog for saving the time of the timer.
In the process, learn how to convert the value of a chronometer.
this is list what i have learned, Dialog setting ,Import image assets.
The process of receiving the value of the chronometer and taking it to the main activity to save it is necessary.
I'm thinking about whether or not to store personal information.